pyproject.toml
[project]
name = "blank-business-builder"
version = "1.0.0"
description = "AI-powered business planning and marketing automation platform with quantum-inspired optimization."
authors = [{ name = "Corporation of Light" }]
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.7",
    "stripe>=11.0.0",
    "aioredis>=2.0.0",
    "celery[redis]>=5.4.0",
    "websockets>=14.0",
    "requests>=2.31.0",
    "aiohttp>=3.10.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "gunicorn>=22.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.13.0",
    "pytest-xdist>=3.6.0",
]

[project.scripts]
bbb = "bbb.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"bbb" = ["*.html"]
